Finalizing NEsper core for 8.5.0

+++ NEsper/NEsper.Common/common/internal/context/controller/condition/ContextControllerConditionFilter.cs
@@ -6,7 +6,6 @@
 // a copy of which has been included with this distribution in the license.txt file.  /
 ///////////////////////////////////////////////////////////////////////////////////////
 
-using System;
 using System.Collections.Generic;
 
 using com.espertech.esper.common.client;
@@ -16,8 +15,6 @@ using com.espertech.esper.common.@internal.context.mgr;
 using com.espertech.esper.common.@internal.context.util;
 using com.espertech.esper.common.@internal.filterspec;
 using com.espertech.esper.common.@internal.filtersvc;
-using com.espertech.esper.compat;
-using com.espertech.esper.compat.collections;
 
 namespace com.espertech.esper.common.@internal.context.controller.condition
 {
@@ -48,9 +45,10 @@ namespace com.espertech.esper.common.@internal.context.controller.condition
 
         public bool Activate(
             EventBean optionalTriggeringEvent,
-            ContextControllerEndConditionMatchEventProvider endConditionMatchEventProvider)
+            ContextControllerEndConditionMatchEventProvider endConditionMatchEventProvider,
+            IDictionary<string, object> optionalTriggeringPattern)
         {
-            AgentInstanceContext agentInstanceContext = controller.Realization.AgentInstanceContextCreate;
+            var agentInstanceContext = controller.Realization.AgentInstanceContextCreate;
 
             FilterHandleCallback filterCallback = new ProxyFilterHandleCallback() {
                 ProcMatchFound = (
@@ -59,27 +57,20 @@ namespace com.espertech.esper.common.@internal.context.controller.condition
                 ProcIsSubselect = () => false,
             };
 
-            FilterValueSetParam[][] addendum = ContextManagerUtil.ComputeAddendumNonStmt(
-                partitionKeys,
-                filter.FilterSpecActivatable,
-                controller.Realization);
             filterHandle = new EPStatementHandleCallbackFilter(
                 agentInstanceContext.EpStatementAgentInstanceHandle,
                 filterCallback);
-            FilterValueSetParam[][] filterValueSet = filter.FilterSpecActivatable.GetValueSet(
-                null,
-                addendum,
-                agentInstanceContext,
-                agentInstanceContext.StatementContextFilterEvalEnv);
-            agentInstanceContext.FilterService.Add(
-                filter.FilterSpecActivatable.FilterForEventType,
-                filterValueSet,
-                filterHandle);
-            long filtersVersion = agentInstanceContext.FilterService.FiltersVersion;
-            agentInstanceContext.EpStatementAgentInstanceHandle.StatementFilterVersion.StmtFilterVersion =
-                filtersVersion;
+            var filterValueSet = ComputeFilterValues(agentInstanceContext);
+            if (filterValueSet != null) {
+                agentInstanceContext.FilterService.Add(
+                    filter.FilterSpecActivatable.FilterForEventType,
+                    filterValueSet,
+                    filterHandle);
+                var filtersVersion = agentInstanceContext.FilterService.FiltersVersion;
+                agentInstanceContext.EpStatementAgentInstanceHandle.StatementFilterVersion.StmtFilterVersion = filtersVersion;
+            }
 
-            bool match = false;
+            var match = false;
             if (optionalTriggeringEvent != null) {
                 match = AgentInstanceUtil.EvaluateFilterForStatement(
                     optionalTriggeringEvent,
@@ -96,24 +87,33 @@ namespace com.espertech.esper.common.@internal.context.controller.condition
                 return;
             }
 
-            AgentInstanceContext agentInstanceContext = controller.Realization.AgentInstanceContextCreate;
-            FilterValueSetParam[][] addendum = ContextManagerUtil.ComputeAddendumNonStmt(
+            var agentInstanceContext = controller.Realization.AgentInstanceContextCreate;
+            var filterValueSet = ComputeFilterValues(agentInstanceContext);
+            if (filterValueSet != null) {
+
+                agentInstanceContext.FilterService.Remove(
+                    filterHandle,
+                    filter.FilterSpecActivatable.FilterForEventType,
+                    filterValueSet);
+                var filtersVersion = agentInstanceContext.StatementContext.FilterService.FiltersVersion;
+                agentInstanceContext.EpStatementAgentInstanceHandle.StatementFilterVersion.StmtFilterVersion =
+                    filtersVersion;
+            }
+
+            filterHandle = null;
+        }
+
+        private FilterValueSetParam[][] ComputeFilterValues(AgentInstanceContext agentInstanceContext)
+        {
+            var addendum = ContextManagerUtil.ComputeAddendumNonStmt(
                 partitionKeys,
                 filter.FilterSpecActivatable,
                 controller.Realization);
-            FilterValueSetParam[][] filterValueSet = filter.FilterSpecActivatable.GetValueSet(
+            return filter.FilterSpecActivatable.GetValueSet(
                 null,
                 addendum,
                 agentInstanceContext,
                 agentInstanceContext.StatementContextFilterEvalEnv);
-            agentInstanceContext.FilterService.Remove(
-                filterHandle,
-                filter.FilterSpecActivatable.FilterForEventType,
-                filterValueSet);
-            filterHandle = null;
-            long filtersVersion = agentInstanceContext.StatementContext.FilterService.FiltersVersion;
-            agentInstanceContext.EpStatementAgentInstanceHandle.StatementFilterVersion.StmtFilterVersion =
-                filtersVersion;
         }
 
         public bool IsImmediate {
@@ -132,10 +132,24 @@ namespace com.espertech.esper.common.@internal.context.controller.condition
             get => null;
         }
 
+
+        public void Transfer(AgentInstanceTransferServices xfer)
+        {
+            if (filterHandle == null) {
+                return;
+            }
+
+            var filterValueSet = ComputeFilterValues(xfer.AgentInstanceContext);
+            if (filterValueSet != null) {
+                xfer.AgentInstanceContext.FilterService.Remove(filterHandle, filter.FilterSpecActivatable.FilterForEventType, filterValueSet);
+                xfer.TargetFilterService.Add(filter.FilterSpecActivatable.FilterForEventType, filterValueSet, filterHandle);
+            }
+        }
+
         private void FilterMatchFound(EventBean theEvent)
         {
             // For OR-type filters we de-duplicate here by keeping the last event instance
-            if (filter.FilterSpecActivatable.Parameters.Length > 1) {
+            if (filter.FilterSpecActivatable.Plan.Paths.Length > 1) {
                 if (theEvent == lastEvent) {
                     return;
                 }
